feat: add user flow documentation and implement Supabase schema initialization
- Created user flow documentation for the application, detailing the overall user journey and specific modes.
- Added Supabase initialization script to set up necessary tables, RLS policies, indexes, and views for user profiles, game history, and achievements.
- Implemented a recovery script for restoring the database schema without dropping existing data.
- Developed components for the versus mode, including RouteCard, VersusBackground, VersusCutins, VersusHud, and the custom hook useVersusLocalGame to manage game state and interactions.
